+++ apps/community/Documents/Loader.tsx
@@ -1,3 +1,5 @@
-import DocumentsTableDocuments from "./Components/TableDocuments"
+import TableDocuments from "./Components/TableDocuments"
+import Browser from "./Components/Browser"
 
-globalThis.main.registerReactComponent('DocumentsTableDocuments', DocumentsTableDocuments);
+globalThis.main.registerReactComponent('DocumentsTableDocuments', TableDocuments);
+globalThis.main.registerReactComponent('DocumentsBrowser', Browser);